Move ProcessResult
