+++ src/Web/ClientApp/src/components/categories/InspectionTypeCategory.tsx
@@ -5,7 +5,7 @@ import { InspectionTypeInteractor } from "../../use-cases";
 import { InspectionTypeController } from "../../controllers";
 import { InspectionTypePresenter } from "../../presenters";
 import { InspectionTypeRepository } from "../../infrastructure/InspectionTypeRepository";
-import { ProcessResult } from "./ProcessResult";
+import { ProcessResult } from "../common/ProcessResult";
 import { BottomNavigationAdd, TopPageLink } from "../common";
 import { EditDialog } from "./EditDialog";
