+++ src/Web/ClientApp/src/components/common/TopPageLink.tsx
@@ -4,4 +4,3 @@ import { Link } from 'react-router-dom';
 export const TopPageLink: FC = (): JSX.Element => {
   return (<Link to="/">トップページへ戻る</Link>);
 }
-
